test(cli): dedupe chatTuiCanStartRootRun table and fold margin-collapse suite into ConversationTranscript
- Delete the standalone chatTuiCanStartRootRun describe in
  chatSessionController.vitest.ts; the TuiStateAndFocus it.each truth
  table plus its session-integrated assertions are a strict superset,
  and the controller delegation test stays. Drop the now-unused import
  specifier.
- Move the five transcript margin-collapse tests from
  TranscriptMarginCollapse.vitest.ts into ConversationTranscript.vitest.ts,
  the suite that already owns transcriptEntryLayout, reusing its
  entry()/phaseRow() fixtures; delete the file.
